backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
django-environ>=0.10,<1.0
django-cors-headers>=3.14,<4.0
drf-yasg>=1.21,<2.0
argon2-cffi>=21.3.0,<22.0
psycopg2-binary>=2.9,<3.0
